backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
torch==2.0.0
opencv-python==4.5.3.56
pillow==8.3.2
numpy==1.21.2
ultralytics==8.0.0